+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpMetricExporter.cs
@@ -75,8 +75,19 @@ public class OtlpMetricExporter : BaseExporter<Metric>
         {
             buffer = this.serializationBuffer.Rent();
 
-            writePosition = ProtobufOtlpMetricSerializer.WriteMetricsData(ref buffer, this.startWritePosition, this.Resource, metrics);
-            serializationSucceeded = true;
+            try
+            {
+                writePosition = ProtobufOtlpMetricSerializer.WriteMetricsData(ref buffer, this.startWritePosition, this.Resource, metrics);
+                serializationSucceeded = true;
+            }
+            catch (Exception ex)
+            {
+                OpenTelemetryProtocolExporterEventSource.Log.BatchDroppedDueToSerializationFailure(
+                    OtlpSignalType.Metrics,
+                    metrics.Count,
+                    ex);
+                return ExportResult.Failure;
+            }
 
             if (this.startWritePosition == GrpcStartWritePosition)
             {