+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -16,6 +16,10 @@ Notes](../../RELEASENOTES.md).
   to be dropped.
   ([#7544](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7544))
 
+* Fixed serialization failures leaving failed trace, metric, and log batches in
+  the serializer's per-thread state. For logs, this also leaked pooled instances.
+  ([#7579](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7579))
+
 ## 1.17.0
 
 Released 2026-Jul-16